# db: add DB.ApplyNoSyncWait for asynchronous apply